Skip updating minimum versions when `!=` is in combined specifier
When `!=` is in the original requirement and is in the supported range,
it can lead to a UnionSpecifier which has `||` as an "or" when you
use `str` on it. This is not standard, and it doesn't look like
dep_logic has a straightforward way of handling it. So, for the time being,
don't bump the minimum versions that should include a `!=`. This might
be a good issue to raise in dep_logic...